requirements.txt
asgiref==3.8.1
Django==5.0.6
django-colorfield==0.11.0
django-jazzmin==3.0.0
django-shortuuidfield==0.1.3
django-taggit==5.0.1
pillow==10.3.0
shortuuid==1.0.13
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
